requirements.txt
arrow==0.17.0
blessed==1.17.11
boto3==1.18.54
cryptography==3.3.1
Django==3.1.3
django-admin-rangefilter==0.8.1
django-cors-headers==3.5.0
django-admin-list-filter-dropdown ==1.0.3
django-cryptography==1.0
django-hashid-field==3.3.1
django-picklefield==3.0.1
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.1
drf-yasg==1.20.0
gunicorn==20.0.4
matplotlib==3.3.3
numpy==1.19.3
pandas==1.1.4
Pillow==8.0.1
psychrochart==0.3.1
PsychroLib==2.5.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.3
requests==2.26.0
scipy==1.5.4